Odev-1/requirements.txt
# Odev-1 — CIFAR-10
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.0.0
tqdm>=4.65.0
scikit-learn>=1.3.0
scikit-image>=0.21.0
matplotlib>=3.7.0
